requirements.txt
ipdb
Django
psycopg2-binary>=2.8
djangorestframework
markdown
django-filter
drf-yasg
django-environ

